pyproject.toml
[project]
name = "reflexion-agent"
version = "0.1.0"
description = "Reflexion Agent with LangGraph - A self-reflection agent implementation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-core>=0.3.19",
    "langgraph>=0.1.0",
    "langgraph-cli[inmem]>=0.1.0",
    "grandalf",
]

authors = [
    {name = "Eden Marco", email = "emarco177@gmail.com"}
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.poetry]
name = "reflexion"
version = "0.1.0"
description = "Reflexion Agent with LangGraph - A self-reflection agent implementation"
authors = ["Eden Marco <emarco177@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "*"
black = "*"
isort = "*"
langchain = "*"
langchain-openai = "*"
grandalf = "*"
langchain-community = "*"
langgraph = "*"
langchain-core = "^0.3.19"

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"
